Fixes #758: Polar Stereographic issues

+++ modules/library/render/src/main/java/org/geotools/renderer/crs/PolarStereographicHandlerFactory.java
@@ -76,9 +76,11 @@ public class PolarStereographicHandlerFactory implements ProjectionHandlerFactor
 
             ReferencedEnvelope validArea;
             if(north) {
-                validArea = new ReferencedEnvelope(-Double.MAX_VALUE, Double.MAX_VALUE, -0, 90, DefaultGeographicCRS.WGS84);
+                validArea = new ReferencedEnvelope(-Integer.MAX_VALUE,
+                        Integer.MAX_VALUE, -0, 90, DefaultGeographicCRS.WGS84);
             } else {
-                validArea = new ReferencedEnvelope(-Double.MAX_VALUE, Double.MAX_VALUE, -90, 0, DefaultGeographicCRS.WGS84);
+                validArea = new ReferencedEnvelope(-Integer.MAX_VALUE,
+                        Integer.MAX_VALUE, -90, 0, DefaultGeographicCRS.WGS84);
             }
             
             return new ProjectionHandler(sourceCrs, validArea, renderingEnvelope) {
